Relocated MantleClientCommand commands to slimeknights.mantle.command.client
